Create FSM_capstone
Worked on this project in a team of 4. FSM class represents the finite state machine to pilot the drone automatically. Implemented this code with help of another teammate.
Followed an Object-Oriented Programming approach to make things easier when editing, as one method represents one state in the state machine.
Methods in this class are called in the main class, where the drone operates by relying on sensors and camera, the latter being regulated by machine learning algorithms.